Create Delete user FE

+++ src/db_querries/delete.js
@@ -7,20 +7,20 @@ const connection = mysql.createConnection({
 });
 
 // Delete user
-const deleteUser = (user_id, callback) => {
-    if (!user_id) {
+const deleteUser = (user_id, user_owner_id, callback) => {
+    if (!user_id || !user_owner_id) {
         callback('id is missed', undefined)
     } else {
         makeDelete();
     }
 
     function makeDelete() {
-        const q = `DELETE FROM friends WHERE id = ${user_id};`;
+        const q = `DELETE FROM friends WHERE user_profile_id = ${user_owner_id} AND id = ${user_id};`;
         connection.query(q, (err, results) => {
             if (err) {
                 callback('No connection to db', undefined);
             } else {
-                callback(undefined, 'user was deleted');
+                callback(undefined, {message: 'user was deleted'});
             }
         });
     }
